Replace old Contact with new Contact resource
