+++ lib/hubspot-ruby.rb
@@ -11,7 +11,6 @@ require 'hubspot/company_properties'
 require 'hubspot/config'
 require 'hubspot/connection'
 require 'hubspot/contact'
-require 'hubspot/contact2'
 require 'hubspot/contact_properties'
 require 'hubspot/contact_list'
 require 'hubspot/form'
