--- spec/fixtures/vcr_cassettes/Hubspot_Contact2_create_with_an_invalid_email_address_raises_an_error.yml
@@ -1,58 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
-    body:
-      encoding: UTF-8
-      string: '{"properties":[{"property":"email","value":"an_invalid_email"}]}'
-    headers:
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 400
-      message: Bad Request
-    headers:
-      Date:
-      - Wed, 27 Feb 2019 15:11:39 GMT
-      Content-Type:
-      - application/json;charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - __cfduid=d7ee8e0eb6f2d454d9a1f58ff60656e6a1551280299; expires=Thu, 27-Feb-20
-        15:11:39 GMT; path=/; domain=.hubapi.com; HttpOnly
-      X-Trace:
-      - 2BEF7D47322E66F4D9070075563590E1349472A4E3000000000000000000
-      X-Hubspot-Ratelimit-Daily:
-      - '160000'
-      X-Hubspot-Ratelimit-Daily-Remaining:
-      - '126825'
-      X-Hubspot-Ratelimit-Secondly:
-      - '60'
-      X-Hubspot-Ratelimit-Secondly-Remaining:
-      - '58'
-      Vary:
-      - Accept-Encoding
-      - Accept-Encoding
-      Access-Control-Allow-Credentials:
-      - 'false'
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Expect-Ct:
-      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
-      Server:
-      - cloudflare
-      Cf-Ray:
-      - 4afb95cdeb91c19a-IAD
-    body:
-      encoding: UTF-8
-      string: '{"validationResults":[{"isValid":false,"message":"Email address an_invalid_email
-        is invalid","error":"INVALID_EMAIL","name":"email"}],"status":"error","message":"Property
-        values were not valid","correlationId":"04bd92c4-baa0-49e6-be8d-fcf53d27c008","requestId":"d3736c642e9c987aa16b3a3d073fb295"}'
-    http_version: 
-  recorded_at: Wed, 27 Feb 2019 15:11:39 GMT
-recorded_with: VCR 4.0.0
