--- spec/fixtures/vcr_cassettes/Hubspot_Contact2_create_with_properties_has_the_property_set.yml
@@ -1,56 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
-    body:
-      encoding: UTF-8
-      string: '{"properties":[{"property":"firstname","value":"Allison"},{"property":"email","value":"iftq80297@example.net"}]}'
-    headers:
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Wed, 27 Feb 2019 15:11:37 GMT
-      Content-Type:
-      - application/json;charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - __cfduid=db9e4cd9f00716fc4828c205a6905d3fe1551280297; expires=Thu, 27-Feb-20
-        15:11:37 GMT; path=/; domain=.hubapi.com; HttpOnly
-      X-Trace:
-      - 2BBFA91CE754BAD3C84BC85760FB6FEB48394951C6000000000000000000
-      X-Hubspot-Ratelimit-Daily:
-      - '160000'
-      X-Hubspot-Ratelimit-Daily-Remaining:
-      - '126829'
-      X-Hubspot-Ratelimit-Secondly:
-      - '60'
-      X-Hubspot-Ratelimit-Secondly-Remaining:
-      - '58'
-      Vary:
-      - Accept-Encoding
-      - Accept-Encoding
-      Access-Control-Allow-Credentials:
-      - 'false'
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Expect-Ct:
-      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
-      Server:
-      - cloudflare
-      Cf-Ray:
-      - 4afb95c41a4d247a-IAD
-    body:
-      encoding: UTF-8
-      string: '{"vid":11601024,"canonical-vid":11601024,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mPEmYrQ-ZvcpNM1tMzLrbvhRuz5nHIXski1-sOnCeWcXf5Sg_KRqnsE6U4yUl8Ox62qyPOokaURluJlD62BMOrD8ZAThCI4YjEQ-MWagnqUw9F8JalJSAF7Ty6f0NbL7ljWS5dc","profile-url":"https://app.hubspot.com/contacts/62515/contact/11601024","properties":{"firstname":{"value":"Allison","versions":[{"value":"Allison","source-type":"API","source-id":null,"source-label":null,"timestamp":1551280297649,"selected":false}]},"email":{"value":"iftq80297@example.net","versions":[{"value":"iftq80297@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1551280297649,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11601024,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"iftq80297@example.net","timestamp":1551280297654,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"01b715d4-e9bf-4195-86e2-e3464de7ec69","timestamp":1551280297664,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
-    http_version: 
-  recorded_at: Wed, 27 Feb 2019 15:11:38 GMT
-recorded_with: VCR 4.0.0
