--- spec/fixtures/vcr_cassettes/Hubspot_Contact2_create_without_properties_creates_a_new_contact.yml
@@ -1,56 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
-    body:
-      encoding: UTF-8
-      string: '{"properties":[{"property":"email","value":"wkde80296@example.net"}]}'
-    headers:
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Wed, 27 Feb 2019 15:11:36 GMT
-      Content-Type:
-      - application/json;charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - __cfduid=ddb48b6535c7e011da6f713175937a6ea1551280296; expires=Thu, 27-Feb-20
-        15:11:36 GMT; path=/; domain=.hubapi.com; HttpOnly
-      X-Trace:
-      - 2BB8493606C98F1236712300EBFBE8D25C867F8C78000000000000000000
-      X-Hubspot-Ratelimit-Daily:
-      - '160000'
-      X-Hubspot-Ratelimit-Daily-Remaining:
-      - '126831'
-      X-Hubspot-Ratelimit-Secondly:
-      - '60'
-      X-Hubspot-Ratelimit-Secondly-Remaining:
-      - '57'
-      Vary:
-      - Accept-Encoding
-      - Accept-Encoding
-      Access-Control-Allow-Credentials:
-      - 'false'
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Expect-Ct:
-      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
-      Server:
-      - cloudflare
-      Cf-Ray:
-      - 4afb95bc3cf323e4-IAD
-    body:
-      encoding: UTF-8
-      string: '{"vid":11600974,"canonical-vid":11600974,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mMORYLBC_COyK05qtqP0AZyRrXgCdJb668WQDCWDYDJBuZyB_Lj8XRBYByO6RAEg9O2hEmOBx6yT2zTyBnVp-0b6dgtBcC775Fzf5ZOfRl3e8whS5PP4rREEWcKfALhR8EkmlKz","profile-url":"https://app.hubspot.com/contacts/62515/contact/11600974","properties":{"email":{"value":"wkde80296@example.net","versions":[{"value":"wkde80296@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1551280296378,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11600974,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"wkde80296@example.net","timestamp":1551280296383,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"2e1f1def-22e4-4ab2-a385-e385100198e4","timestamp":1551280296390,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
-    http_version: 
-  recorded_at: Wed, 27 Feb 2019 15:11:37 GMT
-recorded_with: VCR 4.0.0
