--- spec/fixtures/vcr_cassettes/Hubspot_Contact2_merge_with_valid_contacts_succeeds.yml
@@ -1,162 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
-    body:
-      encoding: UTF-8
-      string: '{"properties":[{"property":"firstname","value":"Rudy"},{"property":"lastname","value":"Hickle"},{"property":"email","value":"31368grhb@example.net"}]}'
-    headers:
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sun, 10 Mar 2019 15:22:48 GMT
-      Content-Type:
-      - application/json;charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - __cfduid=dd88cfa08d31f597b1277abd79c85c1541552231368; expires=Mon, 09-Mar-20
-        15:22:48 GMT; path=/; domain=.hubapi.com; HttpOnly
-      X-Trace:
-      - 2B2AE861C13CB07D1D6DA81184681A1C449F73DA61000000000000000000
-      X-Hubspot-Ratelimit-Daily:
-      - '160000'
-      X-Hubspot-Ratelimit-Daily-Remaining:
-      - '122838'
-      X-Hubspot-Ratelimit-Secondly:
-      - '60'
-      X-Hubspot-Ratelimit-Secondly-Remaining:
-      - '59'
-      Vary:
-      - Accept-Encoding
-      - Accept-Encoding
-      Access-Control-Allow-Credentials:
-      - 'false'
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Expect-Ct:
-      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
-      Server:
-      - cloudflare
-      Cf-Ray:
-      - 4b5649455deacf5a-IAD
-    body:
-      encoding: UTF-8
-      string: '{"vid":11654524,"canonical-vid":11654524,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mPEKlDFZWJtNAC613bhdPnVMIcqaCoqfaFjKYrdCYlwhyyiEu6LWmEbG4mC1bdMZuJgkRXwfXzGc0FkCAZ7D_2YytrGuv4s_GhMDTlDg24RWamJhbsB3P1P7v-BW1nb4w4oGy_t","profile-url":"https://app.hubspot.com/contacts/62515/contact/11654524","properties":{"firstname":{"value":"Rudy","versions":[{"value":"Rudy","source-type":"API","source-id":null,"source-label":null,"timestamp":1552231368561,"selected":false}]},"email":{"value":"31368grhb@example.net","versions":[{"value":"31368grhb@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1552231368561,"selected":false}]},"lastname":{"value":"Hickle","versions":[{"value":"Hickle","source-type":"API","source-id":null,"source-label":null,"timestamp":1552231368561,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11654524,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"31368grhb@example.net","timestamp":1552231368565,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"9782d5f7-88db-4493-8c79-b75f58a274eb","timestamp":1552231368574,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
-    http_version: 
-  recorded_at: Sun, 10 Mar 2019 15:22:48 GMT
-- request:
-    method: post
-    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
-    body:
-      encoding: UTF-8
-      string: '{"properties":[{"property":"firstname","value":"Jacquie"},{"property":"lastname","value":"Ullrich"},{"property":"email","value":"31368vifh@example.net"}]}'
-    headers:
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sun, 10 Mar 2019 15:22:49 GMT
-      Content-Type:
-      - application/json;charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - __cfduid=d721fdfe4b36baaeb2c6de3bee59d17ed1552231369; expires=Mon, 09-Mar-20
-        15:22:49 GMT; path=/; domain=.hubapi.com; HttpOnly
-      X-Trace:
-      - 2B150902A2752886FD7146B1B6CEB2BA9DCEFEFFFB000000000000000000
-      X-Hubspot-Ratelimit-Daily:
-      - '160000'
-      X-Hubspot-Ratelimit-Daily-Remaining:
-      - '122837'
-      X-Hubspot-Ratelimit-Secondly:
-      - '60'
-      X-Hubspot-Ratelimit-Secondly-Remaining:
-      - '59'
-      Vary:
-      - Accept-Encoding
-      - Accept-Encoding
-      Access-Control-Allow-Credentials:
-      - 'false'
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Expect-Ct:
-      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
-      Server:
-      - cloudflare
-      Cf-Ray:
-      - 4b564948ab96cf4e-IAD
-    body:
-      encoding: UTF-8
-      string: '{"vid":11654574,"canonical-vid":11654574,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mOkAc3i1WTaTjrcsU_L3R69CYo-sjgrtp9y5ct7qXNeQKaPzm2SA7MTR02kDtzHk1NP4ZQhkV6RX9IQnk72AApv1bzpJRHzuHoN0ByxgS9CFv-siAy7Jn36jYRbuxxmLS8AUyZJ","profile-url":"https://app.hubspot.com/contacts/62515/contact/11654574","properties":{"firstname":{"value":"Jacquie","versions":[{"value":"Jacquie","source-type":"API","source-id":null,"source-label":null,"timestamp":1552231369098,"selected":false}]},"email":{"value":"31368vifh@example.net","versions":[{"value":"31368vifh@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1552231369098,"selected":false}]},"lastname":{"value":"Ullrich","versions":[{"value":"Ullrich","source-type":"API","source-id":null,"source-label":null,"timestamp":1552231369098,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11654574,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"31368vifh@example.net","timestamp":1552231369103,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"fb3a74e4-c4e7-41c2-9345-c699b76a6c38","timestamp":1552231369115,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
-    http_version: 
-  recorded_at: Sun, 10 Mar 2019 15:22:49 GMT
-- request:
-    method: post
-    uri: https://api.hubapi.com/contacts/v1/contact/merge-vids/11654524/?hapikey=<HAPI_KEY>
-    body:
-      encoding: UTF-8
-      string: '{"vidToMerge":11654574}'
-    headers:
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sun, 10 Mar 2019 15:22:49 GMT
-      Content-Type:
-      - application/json;charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Set-Cookie:
-      - __cfduid=d16308dedab8b3a6651b1ceaff13b6b321552231369; expires=Mon, 09-Mar-20
-        15:22:49 GMT; path=/; domain=.hubapi.com; HttpOnly
-      X-Trace:
-      - 2B4E361147EA620AFAB727AEA521A07F8D00AB7081000000000000000000
-      X-Hubspot-Ratelimit-Daily:
-      - '160000'
-      X-Hubspot-Ratelimit-Daily-Remaining:
-      - '122836'
-      X-Hubspot-Ratelimit-Secondly:
-      - '60'
-      X-Hubspot-Ratelimit-Secondly-Remaining:
-      - '58'
-      Vary:
-      - Accept-Encoding
-      - Accept-Encoding
-      Access-Control-Allow-Credentials:
-      - 'false'
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Expect-Ct:
-      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
-      Server:
-      - cloudflare
-      Cf-Ray:
-      - 4b56494b1917c1a4-IAD
-    body:
-      encoding: UTF-8
-      string: SUCCESS
-    http_version: 
-  recorded_at: Sun, 10 Mar 2019 15:22:49 GMT
-recorded_with: VCR 4.0.0
