+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_create_with_an_existing_email_address_raises_an_error.yml
@@ -1,0 +1,109 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
+    body:
+      encoding: UTF-8
+      string: '{"properties":[{"property":"firstname","value":"Kaylene"},{"property":"lastname","value":"Fahey"},{"property":"email","value":"32035oeuc@example.net"}]}'
+    headers:
+      Content-Type:
+      - application/json
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:33:56 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=dcd07702f43a5c7cf28223af54edac8ec1552232035; expires=Mon, 09-Mar-20
+        15:33:55 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2BBCCD09B8E4C6B15AD21B41899FCED325416EE857000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122569'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '59'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b565990f95dc18b-IAD
+    body:
+      encoding: UTF-8
+      string: '{"vid":11655024,"canonical-vid":11655024,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mNcroHU2Nci8UDAx3HnP93hzPNoxmhX6b5A-uSCivQzvD4_zDeVfqe1xZY0SjpYQnRwTn--koZLphsUZM3m9mYD65RQK2ocLol7ArESKNzH4LTq1upiSyiAWC9G3EEvK9tiAlj5","profile-url":"https://app.hubspot.com/contacts/62515/contact/11655024","properties":{"firstname":{"value":"Kaylene","versions":[{"value":"Kaylene","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232036021,"selected":false}]},"email":{"value":"32035oeuc@example.net","versions":[{"value":"32035oeuc@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232036021,"selected":false}]},"lastname":{"value":"Fahey","versions":[{"value":"Fahey","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232036021,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11655024,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"32035oeuc@example.net","timestamp":1552232036025,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"f695438a-646c-4240-9e4e-e805f99ca352","timestamp":1552232036034,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:33:56 GMT
+- request:
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
+    body:
+      encoding: UTF-8
+      string: '{"properties":[{"property":"email","value":"32035oeuc@example.net"}]}'
+    headers:
+      Content-Type:
+      - application/json
+  response:
+    status:
+      code: 409
+      message: Conflict
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:33:56 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=d93d6dcaa63763bcf2d2e505c63b0962b1552232036; expires=Mon, 09-Mar-20
+        15:33:56 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2B35FA24AE7FD601B73C3918D454CDD937C19099F3000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122568'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '58'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b5659945fa59fc6-IAD
+    body:
+      encoding: UTF-8
+      string: '{"message":"Contact already exists","identityProfile":{"vid":11655024,"identity":[{"value":"32035oeuc@example.net","type":"EMAIL","timestamp":1552232036025,"isPrimary":true},{"value":"f695438a-646c-4240-9e4e-e805f99ca352","type":"LEAD_GUID","timestamp":1552232036034}],"linkedVid":[],"isContact":true,"savedAtTimestamp":1552232036038},"status":"error","correlationId":"f16f3bc9-0f18-4846-b494-11545c16ccb8","error":"CONTACT_EXISTS","requestId":"9da6ce9392d31d95bf0703acd0b69516"}'
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:33:56 GMT
+recorded_with: VCR 4.0.0
