+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_create_with_properties_creates_a_new_contact.yml
@@ -1,0 +1,56 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
+    body:
+      encoding: UTF-8
+      string: '{"properties":[{"property":"firstname","value":"Allison"},{"property":"email","value":"qcau32034@example.net"}]}'
+    headers:
+      Content-Type:
+      - application/json
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:33:54 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=dd3add68220671d03fbe23836afd841f91552232034; expires=Mon, 09-Mar-20
+        15:33:54 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2B19B9CEC4ACBA171A50E0DB17283B95E8D117DD6C000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122572'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '57'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b565988ba7b9586-IAD
+    body:
+      encoding: UTF-8
+      string: '{"vid":11654874,"canonical-vid":11654874,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mNyz_aA4m-gM_dCrOr9ml4NoHxBzXJ48a0kZYsKuNtLuroUKsuMYtfKtW2GQyAl-gohd3VXJCr1yAz7pgLShOO80dwnYMy3AoiSvD-dGYZt-sroEzGA_a-aue1vuLmuheYgizhu","profile-url":"https://app.hubspot.com/contacts/62515/contact/11654874","properties":{"firstname":{"value":"Allison","versions":[{"value":"Allison","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232034707,"selected":false}]},"email":{"value":"qcau32034@example.net","versions":[{"value":"qcau32034@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232034707,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11654874,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"qcau32034@example.net","timestamp":1552232034710,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"079c6a1c-6268-402c-8f09-e79fb0d0d5f9","timestamp":1552232034721,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:33:54 GMT
+recorded_with: VCR 4.0.0
