+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_create_with_properties_is_persisted.yml
@@ -1,0 +1,56 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
+    body:
+      encoding: UTF-8
+      string: '{"properties":[{"property":"firstname","value":"Allison"},{"property":"email","value":"vzpc32035@example.com"}]}'
+    headers:
+      Content-Type:
+      - application/json
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:33:55 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=dc78d34e9bbedf250b24980860c9012c01552232035; expires=Mon, 09-Mar-20
+        15:33:55 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2BB867F9C610A3EDA708B66852AC17DBE96FAC0798000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122570'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '58'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b56598e1e10c1ae-IAD
+    body:
+      encoding: UTF-8
+      string: '{"vid":11654974,"canonical-vid":11654974,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mNbAFNOm-8t669ekch_VQnLD7MMo_Fts4ALleUKpqYK8E0d0YUim4CV_P3574gjnN_Y5Uhq_cxFY2Lh0JeXi0FpkLi3huvQjlc57PARkyEECMhHqgAyO8ShfEcGQkCxmDOr3J-j","profile-url":"https://app.hubspot.com/contacts/62515/contact/11654974","properties":{"firstname":{"value":"Allison","versions":[{"value":"Allison","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232035593,"selected":false}]},"email":{"value":"vzpc32035@example.com","versions":[{"value":"vzpc32035@example.com","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232035593,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11654974,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"vzpc32035@example.com","timestamp":1552232035597,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"47d769c7-f4b9-4459-9b7e-6341a078821b","timestamp":1552232035606,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:33:55 GMT
+recorded_with: VCR 4.0.0
