+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_find_by_email_is_persisted.yml
@@ -1,0 +1,107 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
+    body:
+      encoding: UTF-8
+      string: '{"properties":[{"property":"firstname","value":"Roseline"},{"property":"lastname","value":"Boehm"},{"property":"email","value":"32037hhzb@example.net"}]}'
+    headers:
+      Content-Type:
+      - application/json
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:33:57 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=d9327fb422bad0efa3affb7b4c60a1a521552232037; expires=Mon, 09-Mar-20
+        15:33:57 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2B0DD7D070525D4A29658AC1BAF606B9F14F3271B2000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122564'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '57'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b56599bbc6fc1cc-IAD
+    body:
+      encoding: UTF-8
+      string: '{"vid":11655124,"canonical-vid":11655124,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mPzpKVA_uC-4EXZQ-_KTXoyK4MbLdUVDf-0-V0dru9dKcVqAuvOANNV2_wPK9X8CThhezCsHLAewg_mBiFp3tSG6Ki3stMZIOOzud-3UyO69Gi8zq3PaantkkX42eOebZR1wAwm","profile-url":"https://app.hubspot.com/contacts/62515/contact/11655124","properties":{"firstname":{"value":"Roseline","versions":[{"value":"Roseline","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037743,"selected":false}]},"email":{"value":"32037hhzb@example.net","versions":[{"value":"32037hhzb@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037743,"selected":false}]},"lastname":{"value":"Boehm","versions":[{"value":"Boehm","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037743,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11655124,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"32037hhzb@example.net","timestamp":1552232037747,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"65c16598-0f2f-42a6-a2b4-5001ef2e4257","timestamp":1552232037757,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:33:58 GMT
+- request:
+    method: get
+    uri: https://api.hubapi.com/contacts/v1/contact/email/32037hhzb@example.net/profile?hapikey=<HAPI_KEY>
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers: {}
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:33:58 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=d7a326a427d299dc463ca2356070208851552232038; expires=Mon, 09-Mar-20
+        15:33:58 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2B2AD4DA77083A5F67A6AC4C5B76C85721622AF376000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122563'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '59'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b56599e9bafcf36-IAD
+    body:
+      encoding: UTF-8
+      string: '{"vid":11655124,"canonical-vid":11655124,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mOgmM6zKbCvDfPOIr5oButZW3tlvFbw_Ui1NQaicDH6qlKTam21O_ia-8oHzxEMZLATyoNysbma8V5llI1h-HwMYkkUK84OLxIvVMH0wLkrMl8s3WkxtVR60nb8LaHFjOvrEctw","profile-url":"https://app.hubspot.com/contacts/62515/contact/11655124","properties":{"firstname":{"value":"Roseline","versions":[{"value":"Roseline","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037743,"selected":false}]},"num_conversion_events":{"value":"0","versions":[{"value":"0","source-type":"CALCULATED","source-id":null,"source-label":null,"timestamp":0,"selected":false}]},"lastmodifieddate":{"value":"1552232037815","versions":[{"value":"1552232037815","source-type":"CALCULATED","source-id":null,"source-label":null,"timestamp":1552232037815,"selected":false}]},"num_unique_conversion_events":{"value":"0","versions":[{"value":"0","source-type":"CALCULATED","source-id":null,"source-label":null,"timestamp":0,"selected":false}]},"hs_email_domain":{"value":"example.net","versions":[{"value":"example.net","source-type":"CALCULATED","source-id":null,"source-label":null,"timestamp":1552232037747,"selected":false}]},"createdate":{"value":"1552232037747","versions":[{"value":"1552232037747","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037747,"selected":false}]},"hs_lifecyclestage_subscriber_date":{"value":"1552232037747","versions":[{"value":"1552232037747","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037747,"selected":false}]},"lifecyclestage":{"value":"subscriber","versions":[{"value":"subscriber","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037747,"selected":false}]},"email":{"value":"32037hhzb@example.net","versions":[{"value":"32037hhzb@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037747,"selected":false}]},"lastname":{"value":"Boehm","versions":[{"value":"Boehm","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232037743,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11655124,"saved-at-timestamp":1552232037760,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"32037hhzb@example.net","timestamp":1552232037747,"is-primary":true},{"type":"LEAD_GUID","value":"65c16598-0f2f-42a6-a2b4-5001ef2e4257","timestamp":1552232037757}]}],"merge-audits":[]}'
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:33:58 GMT
+recorded_with: VCR 4.0.0
