+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_find_with_an_invalid_ID_raises_an_error.yml
@@ -2,18 +2,18 @@
 http_interactions:
 - request:
     method: get
-    uri: https://api.hubapi.com/contacts/v1/search/query?count=25&hapikey=<HAPI_KEY>&offset=&q=123xyz
+    uri: https://api.hubapi.com/contacts/v1/contact/vid/0/profile?hapikey=<HAPI_KEY>
     body:
       encoding: US-ASCII
       string: ''
     headers: {}
   response:
     status:
-      code: 200
-      message: OK
+      code: 404
+      message: Not Found
     headers:
       Date:
-      - Sun, 10 Mar 2019 15:09:31 GMT
+      - Sun, 10 Mar 2019 15:33:54 GMT
       Content-Type:
       - application/json;charset=utf-8
       Transfer-Encoding:
@@ -21,14 +21,14 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=dfca0ef3dd4f2308e08da8933da9d01221552230571; expires=Mon, 09-Mar-20
-        15:09:31 GMT; path=/; domain=.hubapi.com; HttpOnly
+      - __cfduid=da8eac2653491846a365818ec470705861552232034; expires=Mon, 09-Mar-20
+        15:33:54 GMT; path=/; domain=.hubapi.com; HttpOnly
       X-Trace:
-      - 2B4DD778A4104F68656A6A85B68F576806A83767E2000000000000000000
+      - 2B329EDE248FEC82BD9BABE1673D60E586383125AA000000000000000000
       X-Hubspot-Ratelimit-Daily:
       - '160000'
       X-Hubspot-Ratelimit-Daily-Remaining:
-      - '123259'
+      - '122574'
       X-Hubspot-Ratelimit-Secondly:
       - '60'
       X-Hubspot-Ratelimit-Secondly-Remaining:
@@ -36,6 +36,8 @@ http_interactions:
       Vary:
       - Accept-Encoding
       - Accept-Encoding
+      X-Hubspot-Notfound:
+      - 'true'
       Access-Control-Allow-Credentials:
       - 'false'
       Strict-Transport-Security:
@@ -45,10 +47,10 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 4b5635cd2b19cf4e-IAD
+      - 4b5659851c499f4e-IAD
     body:
       encoding: UTF-8
-      string: '{"query":"123xyz","offset":0,"has-more":false,"total":0,"contacts":[]}'
+      string: '{"status":"error","message":"contact does not exist","correlationId":"67468b48-e5c0-4510-8dc1-ac9070dba443","requestId":"8c68ade7bde35d445e0f1d0366492872"}'
     http_version: 
-  recorded_at: Sun, 10 Mar 2019 15:09:31 GMT
+  recorded_at: Sun, 10 Mar 2019 15:33:54 GMT
 recorded_with: VCR 4.0.0
