+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_merge_with_invalid_contact_ids_raises_an_error.yml
@@ -1,19 +1,21 @@
 ---
 http_interactions:
 - request:
-    method: get
-    uri: https://api.hubapi.com/contacts/v1/search/query?count=25&hapikey=<HAPI_KEY>&offset=&q=123xyz
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact/merge-vids/1/?hapikey=<HAPI_KEY>
     body:
-      encoding: US-ASCII
-      string: ''
-    headers: {}
+      encoding: UTF-8
+      string: '{"vidToMerge":2}'
+    headers:
+      Content-Type:
+      - application/json
   response:
     status:
-      code: 200
-      message: OK
+      code: 400
+      message: Bad Request
     headers:
       Date:
-      - Sun, 10 Mar 2019 15:07:58 GMT
+      - Sun, 10 Mar 2019 15:34:03 GMT
       Content-Type:
       - application/json;charset=utf-8
       Transfer-Encoding:
@@ -21,14 +23,14 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d2a390928914a34e1f91dc36d7e8148301552230478; expires=Mon, 09-Mar-20
-        15:07:58 GMT; path=/; domain=.hubapi.com; HttpOnly
+      - __cfduid=d8ed5d73369390405ec71548dcd86ad961552232043; expires=Mon, 09-Mar-20
+        15:34:03 GMT; path=/; domain=.hubapi.com; HttpOnly
       X-Trace:
-      - 2B9C654FC869A1EDE0B9829AD43D41557F206CEA95000000000000000000
+      - 2B4297CE0244ED7B2EE200A2A39331B65ED16F615C000000000000000000
       X-Hubspot-Ratelimit-Daily:
       - '160000'
       X-Hubspot-Ratelimit-Daily-Remaining:
-      - '123396'
+      - '122550'
       X-Hubspot-Ratelimit-Secondly:
       - '60'
       X-Hubspot-Ratelimit-Secondly-Remaining:
@@ -45,10 +47,11 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 4b56338c3f33cf62-IAD
+      - 4b5659c0f922cf02-IAD
     body:
       encoding: UTF-8
-      string: '{"query":"123xyz","offset":0,"has-more":false,"total":0,"contacts":[]}'
+      string: '{"status":"error","message":"canonical vid is not a valid contact or
+        is already in the process of being merged","correlationId":"ece817a9-11d5-4bad-bea8-a9c86c7bf562","requestId":"6c9f32062d98ee2ba5c28a1d9273ac38"}'
     http_version: 
-  recorded_at: Sun, 10 Mar 2019 15:07:59 GMT
+  recorded_at: Sun, 10 Mar 2019 15:34:03 GMT
 recorded_with: VCR 4.0.0
