+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_merge_with_valid_contact_ids_succeeds.yml
@@ -1,0 +1,162 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
+    body:
+      encoding: UTF-8
+      string: '{"properties":[{"property":"firstname","value":"Leo"},{"property":"lastname","value":"Toy"},{"property":"email","value":"32040xinp@example.net"}]}'
+    headers:
+      Content-Type:
+      - application/json
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:34:01 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=da439984641ef576a4edac082bdd2b6951552232041; expires=Mon, 09-Mar-20
+        15:34:01 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2BBE6C136C3E972CCD691FF5058C701862767CE4F5000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122554'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '59'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b5659b0898dceb6-IAD
+    body:
+      encoding: UTF-8
+      string: '{"vid":11655174,"canonical-vid":11655174,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mNKejYXV9Agx6OF7c5CgWJ1wegs_x0y6SNFnTI8V9Uq4n76yUVbvtrW3pe0XzNQYOaPADyH5MTFm7STtHXXbEPQ0q9Thc8ikU-E_K3FtUUhA1PpNG6uRXOx7_rUI5P7MWpHvBWz","profile-url":"https://app.hubspot.com/contacts/62515/contact/11655174","properties":{"firstname":{"value":"Leo","versions":[{"value":"Leo","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232041172,"selected":false}]},"email":{"value":"32040xinp@example.net","versions":[{"value":"32040xinp@example.net","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232041172,"selected":false}]},"lastname":{"value":"Toy","versions":[{"value":"Toy","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232041172,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11655174,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"32040xinp@example.net","timestamp":1552232041177,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"fc04c506-3ccb-423c-aaec-31029e2d7297","timestamp":1552232041188,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:34:01 GMT
+- request:
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact?hapikey=<HAPI_KEY>
+    body:
+      encoding: UTF-8
+      string: '{"properties":[{"property":"firstname","value":"Alphonso"},{"property":"lastname","value":"Gutkowski"},{"property":"email","value":"32041adpn@example.com"}]}'
+    headers:
+      Content-Type:
+      - application/json
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:34:02 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=ddfc816dd178141a9d3ee9cae1237a7c11552232042; expires=Mon, 09-Mar-20
+        15:34:02 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2B0ECACFD74C4AC80BB99AF16DDF28759AE180AA6F000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122553'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '59'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b5659bb2ef8cf36-IAD
+    body:
+      encoding: UTF-8
+      string: '{"vid":11655224,"canonical-vid":11655224,"merged-vids":[],"portal-id":62515,"is-contact":true,"profile-token":"AO_T-mNNT1jL68PnTBX5TAhan5Xn2SYiMqEsJr-fHwWR19ioJiq2_FniXW84uOs5YaELF31ZrOTM2FmfpXv_PED-KN0MXTiea_p6bYI7zffbWX9piCBXb8Y6rD9Dpzn4l1HnLuBIpDOU","profile-url":"https://app.hubspot.com/contacts/62515/contact/11655224","properties":{"firstname":{"value":"Alphonso","versions":[{"value":"Alphonso","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232042772,"selected":false}]},"email":{"value":"32041adpn@example.com","versions":[{"value":"32041adpn@example.com","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232042772,"selected":false}]},"lastname":{"value":"Gutkowski","versions":[{"value":"Gutkowski","source-type":"API","source-id":null,"source-label":null,"timestamp":1552232042772,"selected":false}]}},"form-submissions":[],"list-memberships":[],"identity-profiles":[{"vid":11655224,"is-deleted":false,"is-contact":false,"pointer-vid":0,"previous-vid":0,"linked-vids":[],"saved-at-timestamp":0,"deleted-changed-timestamp":0,"identities":[{"type":"EMAIL","value":"32041adpn@example.com","timestamp":1552232042775,"is-primary":true,"source":"UNSPECIFIED"},{"type":"LEAD_GUID","value":"0bae6712-68a9-42ce-9d84-4769e2a06e88","timestamp":1552232042785,"source":"UNSPECIFIED"}]}],"merge-audits":[]}'
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:34:03 GMT
+- request:
+    method: post
+    uri: https://api.hubapi.com/contacts/v1/contact/merge-vids/11655174/?hapikey=<HAPI_KEY>
+    body:
+      encoding: UTF-8
+      string: '{"vidToMerge":11655224}'
+    headers:
+      Content-Type:
+      - application/json
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Sun, 10 Mar 2019 15:34:03 GMT
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Set-Cookie:
+      - __cfduid=da9f5e2083feb85f633a29f7e930273ac1552232043; expires=Mon, 09-Mar-20
+        15:34:03 GMT; path=/; domain=.hubapi.com; HttpOnly
+      X-Trace:
+      - 2B8521F23BCD49632C2B2AC2F28AF3B6EF5D95B4E5000000000000000000
+      X-Hubspot-Ratelimit-Daily:
+      - '160000'
+      X-Hubspot-Ratelimit-Daily-Remaining:
+      - '122552'
+      X-Hubspot-Ratelimit-Secondly:
+      - '60'
+      X-Hubspot-Ratelimit-Secondly-Remaining:
+      - '59'
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      Access-Control-Allow-Credentials:
+      - 'false'
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 4b5659be7a14c1c2-IAD
+    body:
+      encoding: UTF-8
+      string: SUCCESS
+    http_version: 
+  recorded_at: Sun, 10 Mar 2019 15:34:03 GMT
+recorded_with: VCR 4.0.0
