+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_search_when_the_query_returns_no_contacts_does_not_have_a_next_page.yml
@@ -13,7 +13,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Sun, 10 Mar 2019 15:08:39 GMT
+      - Sun, 10 Mar 2019 15:34:00 GMT
       Content-Type:
       - application/json;charset=utf-8
       Transfer-Encoding:
@@ -21,18 +21,18 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=dee8864b133ba3ba1e5e97f6b92fd9d981552230519; expires=Mon, 09-Mar-20
-        15:08:39 GMT; path=/; domain=.hubapi.com; HttpOnly
+      - __cfduid=d507ed14af3f8a5f1146bde54b4da1eb41552232040; expires=Mon, 09-Mar-20
+        15:34:00 GMT; path=/; domain=.hubapi.com; HttpOnly
       X-Trace:
-      - 2B3502D6D38DDC4750B38AC3F391516B419801BC37000000000000000000
+      - 2B14DE045B80A8723371519FA114E27978ED29B77C000000000000000000
       X-Hubspot-Ratelimit-Daily:
       - '160000'
       X-Hubspot-Ratelimit-Daily-Remaining:
-      - '123299'
+      - '122555'
       X-Hubspot-Ratelimit-Secondly:
       - '60'
       X-Hubspot-Ratelimit-Secondly-Remaining:
-      - '59'
+      - '58'
       Vary:
       - Accept-Encoding
       - Accept-Encoding
@@ -45,10 +45,10 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 4b563487c9aacf2a-IAD
+      - 4b5659af3f69cf5a-IAD
     body:
       encoding: UTF-8
       string: '{"query":"123xyz","offset":0,"has-more":false,"total":0,"contacts":[]}'
     http_version: 
-  recorded_at: Sun, 10 Mar 2019 15:08:39 GMT
+  recorded_at: Sun, 10 Mar 2019 15:34:00 GMT
 recorded_with: VCR 4.0.0
