+++ spec/fixtures/vcr_cassettes/Hubspot_Contact_search_when_the_query_returns_no_contacts_has_no_contacts.yml
@@ -2,18 +2,18 @@
 http_interactions:
 - request:
     method: get
-    uri: https://api.hubapi.com/contacts/v1/contact/vid/0/profile?hapikey=<HAPI_KEY>
+    uri: https://api.hubapi.com/contacts/v1/search/query?count=25&hapikey=<HAPI_KEY>&offset=&q=123xyz
     body:
       encoding: US-ASCII
       string: ''
     headers: {}
   response:
     status:
-      code: 404
-      message: Not Found
+      code: 200
+      message: OK
     headers:
       Date:
-      - Wed, 27 Feb 2019 15:11:36 GMT
+      - Sun, 10 Mar 2019 15:33:58 GMT
       Content-Type:
       - application/json;charset=utf-8
       Transfer-Encoding:
@@ -21,23 +21,21 @@ http_interactions:
       Connection:
       - keep-alive
       Set-Cookie:
-      - __cfduid=d091b7eb2ed3307535158a5ed833e07161551280296; expires=Thu, 27-Feb-20
-        15:11:36 GMT; path=/; domain=.hubapi.com; HttpOnly
+      - __cfduid=d230fc50e90f8723e97397ecbaea6ad951552232038; expires=Mon, 09-Mar-20
+        15:33:58 GMT; path=/; domain=.hubapi.com; HttpOnly
       X-Trace:
-      - 2B0193AB18400D5B828A1F130A333E1DA3512E2BEE000000000000000000
+      - 2B38DD89DBB8BC39A3A9057A1E5705F83A5CA32632000000000000000000
       X-Hubspot-Ratelimit-Daily:
       - '160000'
       X-Hubspot-Ratelimit-Daily-Remaining:
-      - '126832'
+      - '122561'
       X-Hubspot-Ratelimit-Secondly:
       - '60'
       X-Hubspot-Ratelimit-Secondly-Remaining:
-      - '58'
+      - '57'
       Vary:
       - Accept-Encoding
       - Accept-Encoding
-      X-Hubspot-Notfound:
-      - 'true'
       Access-Control-Allow-Credentials:
       - 'false'
       Strict-Transport-Security:
@@ -47,10 +45,10 @@ http_interactions:
       Server:
       - cloudflare
       Cf-Ray:
-      - 4afb95bb8e49c1f4-IAD
+      - 4b5659a2fe15cf02-IAD
     body:
       encoding: UTF-8
-      string: '{"status":"error","message":"contact does not exist","correlationId":"d32dca4b-b535-497b-86ee-458d1ff27d75","requestId":"b5c96bf7fb8536c50184a0aeb0d42cfc"}'
+      string: '{"query":"123xyz","offset":0,"has-more":false,"total":0,"contacts":[]}'
     http_version: 
-  recorded_at: Wed, 27 Feb 2019 15:11:36 GMT
+  recorded_at: Sun, 10 Mar 2019 15:33:58 GMT
 recorded_with: VCR 4.0.0
