+++ spec/lib/hubspot/engagement_spec.rb
@@ -75,12 +75,12 @@ describe Hubspot::Engagement do
       cassette "engagement_associate"
 
       let(:engagement) { Hubspot::EngagementNote.create!(nil, 'note') }
-      let(:contact) { Hubspot::Contact.create!("newcontact#{Time.now.to_i}@hsgem.com") }
-      subject { Hubspot::Engagement.associate!(engagement.id, 'contact', contact.vid) }
+      let(:contact) { Hubspot::Contact.create(email: "newcontact#{Time.now.to_i}@hsgem.com") }
+      subject { Hubspot::Engagement.associate!(engagement.id, 'contact', contact.id) }
 
       it 'associate an engagement to a resource' do
         subject
-        found_by_contact = Hubspot::Engagement.find_by_contact(contact.vid)
+        found_by_contact = Hubspot::Engagement.find_by_contact(contact.id)
         expect(found_by_contact.first.id).to eql engagement.id
       end
     end
